move TokenParser logic in class Md
